fallback to english if no comments in current language fixes #421 (#431)
Fallback to english if no comments in current language fixes #421. Only fallback to english if there are english comments. The select contains the en (english) option only if there are english comments available, so we can do the check client side and only hit the server if there comments are available.