Update CRL due to certificate overlap
There was an issue in our signing server deployment so that some certificates got the same serial number. If one of them is now revoked the other one with the same serial number is also revoked. So we decided to re-issue all affected certificates. Old ones should still work during the installation, but you need the new certificate when uploading a new version of the app to the app store.

Signed-off-by: Morris Jobke <hey@morrisjobke.de>